GH-472 Fix the row/col mixup

+++ NeoVimServer/server_ui.m
@@ -674,10 +674,10 @@ static NeoVimBuffer *buffer_for(buf_T *buf) {
 void neovim_scroll(void **argv) {
   work_and_write_data_sync(argv, ^NSData *(NSData *data) {
     NSInteger *values = (NSInteger *) data.bytes;
-    NSInteger horiz = values[0];
-    NSInteger vert = values[1];
-    NSInteger row = values[2];
-    NSInteger column = values[3];
+    int horiz = (int) values[0];
+    int vert = (int) values[1];
+    int row = (int) values[2];
+    int column = (int) values[3];
 
     if (horiz == 0 && vert == 0) {
       return nil;
@@ -693,11 +693,11 @@ void neovim_scroll(void **argv) {
     int vertDir;
     if (horiz != 0) {
       horizDir = horiz > 0 ? MSCR_RIGHT: MSCR_LEFT;
-      custom_ui_scroll(horizDir, (int) ABS(horiz), row, column);
+      custom_ui_scroll(horizDir, ABS(horiz), row, column);
     }
     if (vert != 0) {
       vertDir = vert > 0 ? MSCR_DOWN: MSCR_UP;
-      custom_ui_scroll(vertDir, (int) ABS(vert), row, column);
+      custom_ui_scroll(vertDir, ABS(vert), row, column);
     }
 
     refresh_ui_screen(VALID);
